Fix window borders showing while gaps are disabled
While gaps are disabled in monocle layout or tiling layout with one
window, borders would show on adjacent monitors.

The commit makes it so that the border width is set to 0 if in monocle
layout or there is only one tiling window. Offsets and window sizes are
adjusted in resize().